refactored WebSocket Service and Client
