+++ space-client/src/app/app.module.ts
@@ -12,9 +12,9 @@ import { AppState, InternalStateType } from './app.service';
 import { NoContentComponent } from './no-content';
 import { DebugClientComponent } from './debug-client/debug-client.component';
 import { SpaceDebugHttpService } from './debug-client/space-debug-http.service';
-import { SpaceWebSocketService } from './ws-client/space-web-socket.service';
+import { SpaceWebSocketService } from './web-socket-client/service/space-web-socket.service';
 import { WebSocketService } from './shared/services/websocket.service';
-import { WsClientComponent } from './ws-client/ws-client.component';
+import { WebSocketClientComponent } from './web-socket-client/web-socket-client.component';
 
 // Application wide providers
 const APP_PROVIDERS = [
@@ -37,7 +37,7 @@ type StoreType = {
     AppComponent,
     NoContentComponent,
     DebugClientComponent,
-    WsClientComponent
+    WebSocketClientComponent
   ],
   imports: [ // import Angular's modules
     BrowserModule,
