+++ space-client/src/app/web-socket-client/web-socket-client.component.css
